Feature: Use context for contact modal.
