+++ components/contactModal/contactModal.module.scss
@@ -1,0 +1,23 @@
+@import "../../styles/global.module.scss";
+
+.contactPopup {
+  text-align: center;
+  min-height: 20em;
+  padding: 2em;
+  background: $white;
+  .popupTitle {
+    color: $black;
+  }
+  .emailLink {
+    color: $black;
+    font-family: $body-font-family;
+    font-weight: normal;
+    text-decoration: none;
+    &:hover {
+      color: $brand-primary;
+    }
+  }
+  img { 
+    width: 20em;
+  }
+}
